backend/requirements.txt
django==6.0.3
djangorestframework==3.16.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
python-dotenv==1.1.1
